+++ components/render/render.h
@@ -49,8 +49,8 @@ namespace Render
     class SpriteCacheBase
     {
         public:
-            virtual SpriteGroup* get(size_t key) = 0;
-            virtual void setImmortal(size_t index, bool immortal) = 0;
+            virtual SpriteGroup* get(uint32_t key) = 0;
+            virtual void setImmortal(uint32_t index, bool immortal) = 0;
     };
 
 
@@ -68,7 +68,7 @@ namespace Render
     void quitGui();
     void drawGui(std::vector<DrawCommand>& buffer, SpriteCacheBase* cache);
 
-    bool getImageInfo(const std::string& path, size_t& width, size_t& height, size_t& animLength, int32_t celIndex=0);
+    bool getImageInfo(const std::string& path, uint32_t& width, uint32_t& height, uint32_t& animLength, int32_t celIndex=0);
     void drawCursor(Sprite s, size_t w=0, size_t h=0);
     SpriteGroup* loadSprite(const std::string& path, bool hasTrans, size_t transR, size_t transG, size_t transB);
     SpriteGroup* loadVanimSprite(const std::string& path, size_t vAnim, bool hasTrans, size_t transR, size_t transG, size_t transB);
